Add sample code TSI_KS_OTP_Verify for OTP key verification
This sample demonstrates how to verify the correctness of OTP keys
using the Crypto AES and ECC engines. Users must know the expected
content of the OTP keys in advance and fill them into the
`my_opt_keys[]` array in the code.

The verification process works by performing AES/ECC operations
separately on both the known plaintext and the OTP keys. If the
outputs match, it confirms that the programmed OTP key is identical
to the declared plaintext.

Change-Id: Id1e238a4c859f0b34fde2c7b404d76fcf1b5005e